[Developing] Include condition classification
This commit mainly added the condition classification
functionality to the project.

Based on the condition attribute labels that are given
in the 100k package, three resnet18 models are trained
to classify new images to different weather, scene and
timeofday labels.

The condition labels of 10k packages are generated by
these models, and the performance evaluation of models
are logged under /output/classification.